Using uint32_t for less chance of overflow in activityClassifier.c.

+++ src/boot/ksdk1.1.0/activityClassifier.c
@@ -84,7 +84,7 @@ uint32_t sqrtInt(uint32_t base){
 WarpStatus updateAccelerations(){
   uint16_t XLSB, YLSB, ZLSB; // Least significant byte of each acceleration measurement.
   uint16_t XMSB, YMSB, ZMSB; // Most significant byte of each acceleration measurement.
-  float XAcceleration, YAcceleration, ZAcceleration; // Actual acceleration values for checking their accuracy.
+  uint32_t XAcceleration, YAcceleration, ZAcceleration; // Actual acceleration values for checking their accuracy.
   WarpStatus i2cReadStatus;
 
   warpScaleSupplyVoltage(deviceMMA8451QState.operatingVoltageMillivolts);
@@ -105,7 +105,7 @@ WarpStatus updateAccelerations(){
   XLSB = deviceMMA8451QState.i2cBuffer[1];
   XCombined = ((XMSB & 0xFF) << 6) | (XLSB >> 2);
   XCombined = (XCombined ^ (1 << 13)) - (1 << 13);
-  XAcceleration = XCombined * 4787; // Convert the acceleration to ums^-2.
+  XAcceleration = (uint32_t)XCombined * 4787; // Convert the acceleration to ums^-2.
   warpPrint("XMSB: %d.\n", XMSB);
   warpPrint("XLSB: %d.\n", XLSB);
   warpPrint("XCombined - Decimal: %d, Hexadecimal: %x.\n", XCombined, XCombined);
@@ -115,7 +115,7 @@ WarpStatus updateAccelerations(){
   YLSB = deviceMMA8451QState.i2cBuffer[3];
   YCombined = ((YMSB & 0xFF) << 6) | (YLSB >> 2);
   YCombined = (YCombined ^ (1 << 13)) - (1 << 13);
-  YAcceleration = YCombined * 4787; // Convert the acceleration to ums^-2.
+  YAcceleration = (uint32_t)YCombined * 4787; // Convert the acceleration to ums^-2.
   // warpPrint("YMSB: %d.\n", YMSB);
   // warpPrint("YLSB: %d.\n", YLSB);
   // warpPrint("YCombined - Decimal: %d, Hexadecimal: %x.\n", YCombined, YCombined);
@@ -125,13 +125,13 @@ WarpStatus updateAccelerations(){
   ZLSB = deviceMMA8451QState.i2cBuffer[5];
   ZCombined = ((ZMSB & 0xFF) << 6) | (ZLSB >> 2);
   ZCombined = (ZCombined ^ (1 << 13)) - (1 << 13);
-  ZAcceleration = ZCombined * 4787; // Convert the acceleration to ums^-2.
+  ZAcceleration = (uint32_t)ZCombined * 4787; // Convert the acceleration to ums^-2.
   // warpPrint("ZMSB: %d.\n", ZMSB);
   // warpPrint("ZLSB: %d.\n", ZLSB);
   // warpPrint("ZCombined - Decimal: %d, Hexadecimal: %x.\n", ZCombined, ZCombined);
   // warpPrint("ZAcceleration (mms^-2) - Decimal: %d, Hexadecimal: %x.\n", ZAcceleration / 1000, ZAcceleration / 1000);
 
-  accelerationMagnitude = sqrtInt((uint32_t)(XAcceleration*XAcceleration) + (uint32_t)(YAcceleration*YAcceleration) + (uint32_t)(ZAcceleration*ZAcceleration));
+  accelerationMagnitude = sqrtInt((XAcceleration*XAcceleration) + (YAcceleration*YAcceleration) + (ZAcceleration*ZAcceleration));
 
   // Shift AccelerationBuffer and LPFBuffer left to free up space for new data.
   for (int i = 1; i < BUFFER_SIZE; i++){
